examples/backend/pyproject.toml
[project]
name = "allauth-js-example-backend"
version = "0.1.0"
description = "Minimal Django backend example for allauth-js library"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1,<5.2",
    "django-allauth[socialaccount]>=65.0,<66.0",
    "django-cors-headers>=4.5,<5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config"]
